--- src/components/templates/Template.test.tsx
@@ -1,19 +1,0 @@
-import { render, screen } from "@testing-library/react";
-import React from "react";
-
-import Template from "./Template";
-
-describe("Templateコンポーネントのテスト", () => {
-  const headerContent = <div>Header Content</div>;
-  const mainContent = <div>Main Content</div>;
-
-  test("ヘッダーが正しく表示されること", () => {
-    render(<Template content={mainContent} header={headerContent} />);
-    expect(screen.getByText("Header Content")).toBeInTheDocument();
-  });
-
-  test("メインコンテンツが正しく表示されること", () => {
-    render(<Template content={mainContent} header={headerContent} />);
-    expect(screen.getByText("Main Content")).toBeInTheDocument();
-  });
-});
